requirements.txt
asgiref==3.11.0
Django==5.2.8
pillow==12.0.0
sqlparse==0.5.3
openai==1.51.0
pytesseract==0.3.13
pandas==2.2.2
numpy==2.0.1
requests==2.32.3
